Multiple changes
- Removed some warnings.
- Added OGLPLUS_ABORT.

+++ include/oglplus/dsa/ext/framebuffer.hpp
@@ -465,7 +465,7 @@ public:
 	{
 		OGLPLUS_GLFUNC(FramebufferDrawBuffersEXT)(
 			_obj_name(),
-			buffers.Count(),
+			GLint(buffers.Count()),
 			buffers.Values()
 		);
 		OGLPLUS_VERIFY(
